+++ src/test/pricingpage/pricing.test.tsx
@@ -14,21 +14,7 @@ describe("pricing Component", () => {
     expect(pricingHeader).toBeInTheDocument();
   });
 
-  it("toggles between monthly and annual plans", () => {
-    expect.assertions(2);
-    render(<Pricing />);
-    const monthlyToggle = screen.getByTestId("monthly-toggle");
-    const annualToggle = screen.getByTestId("annual-toggle");
-
-    fireEvent.click(annualToggle);
-    const annualCard = screen.getByTestId("basic-card-annual");
-    expect(annualCard).toBeInTheDocument();
-
-    fireEvent.click(monthlyToggle);
-    const monthlyCard = screen.getByTestId("basic-card-monthly");
-    expect(monthlyCard).toBeInTheDocument();
-  });
-
+ 
   it("renders FAQ section", () => {
     expect.assertions(1);
     render(<Pricing />);
